refactor(architecture): FSD `domain` 레이어 구축 및 비즈니스 로직 그룹화
FSD 아키텍처 리팩토링 2단계로, 기존 `stores`, `hooks`, `utils`에 분산되어 있던 핵심 비즈니스 로직을 `domain` 레이어로 통합하여 재구성합니다.

이를 통해 데이터 모델과 관련된 상태, 훅, 유틸리티의 응집도를 극대화하고, 코드의 예측 가능성을 높여 유지보수성을 향상시킵니다.

**주요 변경 사항:**
- `src/domain` 디렉토리 생성 및 4개의 핵심 도메인(`dashboard`, `analytics`, `youtube`, `platform`) 구조 설계
- 총 9개의 파일(store, hook, util, constant)을 각 도메인의 `model` 세그먼트로 이동
- `page_store`, `notification_store` 등 소속이 불분명한 전역 상태는 의도적으로 기존 위치에 유지하여 다음 단계에서 처리하도록 분리
- 파일 이동에 따라 20개 이상 파일의 `import` 경로를 새로운 `domain` 경로로 업데이트
