+++ src/common/ui/AudienceDemographicsChart.jsx
@@ -1,7 +1,7 @@
 import { BarChart, Bar, XAxis, YAxis, CartesianGrid, Tooltip, Legend, ResponsiveContainer } from 'recharts';
 import { motion } from 'framer-motion';
 import { Users } from 'lucide-react';
-import { audience_demo_data } from '../../utils/dashboard_constants.js';
+import { audience_demo_data } from '../../domain/dashboard/model/dashboardConstants.js';
 import GlassCard from './glass-card.jsx';
 
 const AudienceDemographicsChart = () => {
