+++ src/components/dashboard/dashboard_header.jsx
@@ -7,7 +7,7 @@ import React, { useCallback } from 'react';
 import { motion } from 'framer-motion';
 import { Sun, Moon } from 'lucide-react';
 import { usePageStore } from '../../stores/page_store.js';
-import { get_header_info } from '../../utils/dashboard_utils.js';
+import { get_header_info } from '../../domain/dashboard/model/dashboardUtils.js';
 import GlassCard from '../../common/ui/glass-card.jsx';
 import Notification from '../../common/ui/notification.jsx';
