+++ src/components/dashboard/settings_view.jsx
@@ -6,7 +6,7 @@
 import React from 'react';
 import ConnectionManagementCard from "./connection-management-card.jsx";
 import { DataExportCard } from "./data-export-card.jsx";
-import { get_platform_data } from '../../utils/dashboard_utils.js';
+import { get_platform_data } from '../../domain/dashboard/model/dashboardUtils.js';
 
 /**
  * Settings View 컴포넌트
