+++ src/domain/youtube/model/useYouTubeChannelInfo.js
@@ -1,7 +1,7 @@
 import { useEffect, useState, useCallback } from 'react';
-import { getYouTubeChannelId } from '../common/api/api.js';
-import { useYouTubeStore } from '../stores/youtube_store.js';
-import { usePlatformStore } from '../stores/platform_store.js';
+import { getYouTubeChannelId } from '../../../common/api/api.js';
+import { useYouTubeStore } from './store.js';
+import { usePlatformStore } from '../../platform/model/store.js';
 
 export const useYouTubeChannelInfo = () => {
   const { setChannelInfo } = useYouTubeStore();
